Finalized transition to using user subscriptions.
For a payment an order is generated. This order's id is then used for generating a payment after it has been performed by the user. If it is paid it's id is added to a user's 'payment' list. Generated orders maintain references to the user that initiated it in order to allow the callback to associate the payment with the user.