chore(invoice): re-structuring Invoice type.
Add updating invoices and tests for them as they were missing.
Change endpoint for listing invoices to take a locked-parameter.
Fix the entire package not considering a list of recipients (earlier only one recipient).

Instead of not being able to send locked invoices, only able to send pending invoices.
Update the Invoice type to use a union of LockedInvoice and UnlockedInvoice.

Update types to use Array instead of [] for uniformity.

Related to #94.